requirements.txt
pandas~=2.2.0
numpy~=1.26.4
matplotlib~=3.8.2
bs4~=0.0.2
beautifulsoup4~=4.12.3
dash~=2.15.0
pyyaml~=6.0.1
Flask~=3.0.2
plotly~=5.18.0
dash-bootstrap-components~=1.5.0
dash-bootstrap-templates~=1.1.2